Progress, we haz it.
We can now do a lot more different types of questions (added variable operators), we have levels (although I'm not entirely happy with how we provide them) and we keep track of all questions answered within a level. We also have a way to end a level (but it's not really sexy yet).

Lots to do, but a lot of progress as well.